+++ CHANGELOG.md
@@ -17,6 +17,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   feature flag).
 - Added `bytes` module guarded by the `bytes` feature, containing
   `constant_time_eq`.
+- `hmac::Hmac` now implements `Clone` and `std::hash::Hasher`, allowing it to be
+  used with any type that implements `std::hash::Hash`.
 
 ### Changed
 - `build.rs` implements symbol name scraping natively, and no longer relies on